修整一下代码，准备编写分布式的node和coordinate模块。 已经有了假想的分布式结构和rpc调用过程

+++ core/types.h
@@ -17,8 +17,8 @@ namespace gsf
 	typedef uint16_t SessionID;
 	static const uint16_t SessionNil = 0;
 
-	typedef uint16_t MsgID;
-	static const uint16_t MsgNil = 0;
+	typedef int32_t MsgID;
+	static const int32_t MsgNil = 0;
 
 	static const uint16_t LogErr = 0;
 	static const uint16_t LogWarning = 1;
